+++ precompiles/ics20/tx.go
@@ -47,32 +47,9 @@ func (p Precompile) Transfer(
 		return nil, err
 	}
 
-	// no need to have authorization when the contract caller is the same as origin (owner of funds)
-	// and the sender is the origin
-<<<<<<< HEAD
-	var (
-		expiration *time.Time
-		auth       authz.Authorization
-		resp       *authz.AcceptResponse
-	)
-
-	if contract.CallerAddress != origin {
-		// check if authorization exists
-		auth, expiration, err = authorization.CheckAuthzExists(ctx, p.AuthzKeeper, contract.CallerAddress, origin, TransferMsg)
-		if err != nil {
-			return nil, fmt.Errorf(authorization.ErrAuthzDoesNotExistOrExpired, contract.CallerAddress, origin)
-		}
-
-		// Accept the grant and return an error if the grant is not accepted
-		resp, err = p.AcceptGrant(ctx, contract.CallerAddress, origin, msg, auth)
-		if err != nil {
-			return nil, err
-		}
-=======
 	resp, expiration, err := CheckAndAcceptAuthorizationIfNeeded(ctx, contract, origin, p.AuthzKeeper, msg)
 	if err != nil {
 		return nil, err
->>>>>>> 6d2d0f1f (fix(ics20): Extract grant checking and updating functions for reuse (#1850))
 	}
 
 	res, err := p.transferKeeper.Transfer(sdk.WrapSDKContext(ctx), msg)
@@ -80,17 +57,8 @@ func (p Precompile) Transfer(
 		return nil, err
 	}
 
-<<<<<<< HEAD
-	// Update grant only if is needed
-	if contract.CallerAddress != origin {
-		// accepts and updates the grant adjusting the spending limit
-		if err = p.UpdateGrant(ctx, contract.CallerAddress, origin, expiration, resp); err != nil {
-			return nil, err
-		}
-=======
 	if err := UpdateGrantIfNeeded(ctx, contract, p.AuthzKeeper, origin, expiration, resp); err != nil {
 		return nil, err
->>>>>>> 6d2d0f1f (fix(ics20): Extract grant checking and updating functions for reuse (#1850))
 	}
 
 	if err = p.EmitIBCTransferEvent(